refactor(cli): move type-extractor-only detectors out of shared
branded-type-detector, interface-detector, ignore-fields-detector/
validator, and inline-object-extractor/utils had exactly one consumer
(type-extractor) despite living in shared/. Move them into a new
type-extractor/detector/ directory, mirroring the stage's existing
extractor/converter/collector/mapper/validator layout.

Also split shared/metadata-detector.ts: getActualMetadataType is a
generic optional-property unwrap helper reused by several shared
modules and by resolver-extractor, so it stays in shared/. The scalar-
classification algorithm (detectScalarMetadata and friends) is used
only by type-extractor/extractor/{type-extractor,field-type-resolver}.ts,
so it moves next to its consumers as
type-extractor/extractor/scalar-metadata-detector.ts. Function bodies
are preserved as-is; only import paths change.
