+++ packages/cli/src/type-extractor/detector/ignore-fields-validator.test.ts
