+++ packages/cli/src/type-extractor/detector/ignore-fields-validator.ts
@@ -6,7 +6,7 @@
  * all fields are excluded.
  */
 
-import type { Diagnostic, SourceLocation } from "../core/index.js";
+import type { Diagnostic, SourceLocation } from "../../core/index.js";
 
 /**
  * Parameters for validateIgnoreFields function.
