+++ packages/cli/src/type-extractor/detector/inline-object-utils.ts
